test: move repro to summarizeWithOutOfOrderDataStoreRealization and assert current behavior
Move the summarizer realize-between-summarize-and-refresh repro next to the similar out-of-order realization tests. The summary upload currently fails in this scenario; the test asserts that (assert.rejects) and documents that we intentionally do not fix it since the summarizer restarts and self-heals. Wrapped in itExpects to allow the expected Summarize_cancel/SummarizeFailed error events.
